Add: Products models

+++ RSMEnterpriseIntegrationsAPI/Application/DTOs/CreateProductDto.cs
@@ -6,8 +6,8 @@ namespace RSMEnterpriseIntegrationsAPI.Application.DTOs
         public string ProductNumber { get; set; } = string.Empty;
         public bool MakeFlag { get; set; }
         public bool FinishedGoodsFlag { get; set; }
-        public int SafetyStockLevel { get; set; }
-        public int ReorderPoint { get; set; }
+        public short SafetyStockLevel { get; set; }
+        public short ReorderPoint { get; set; }
         public decimal StandardCost { get; set; }
         public decimal ListPrice { get; set; }
         public DateTime SellStartDate { get; set; }
